rename: automation campaign history to campaign history
