+++ src/Enums/CommunicationHistoryStatus.php
@@ -1,8 +1,10 @@
 <?php
 
+declare(strict_types=1);
+
 namespace YouNegotiate\Enums;
 
-enum AutomationCampaignHistoryStatus: int
+enum CommunicationHistoryStatus: int
 {
     case SUCCESS = 1;
     case FAILED = 2;
